+++ lib/urlwatch/jobs.py
@@ -180,7 +180,7 @@ class JobBase(object, metaclass=TrackSubClasses):
 
 class Job(JobBase):
     __required__ = ()
-    __optional__ = ('name', 'filter', 'max_tries', 'diff_tool', 'compared_versions', 'diff_filter')
+    __optional__ = ('name', 'filter', 'max_tries', 'diff_tool', 'compared_versions', 'diff_filter', 'treat_new_as_changed')
 
     # determine if hyperlink "a" tag is used in HtmlReporter
     LOCATION_IS_URL = False